Update README.md
The objective of this project is to familiarize you with the creation and execution of threads using the Thread class methods. You should use, when necessary, the following methods to synchronize all threads: run(), start(), currentThread(), getName(), join(), yield(), sleep(time), isAlive(), getPriority(), setPriority(), interrupt(), isInterrupted().

The use of semaphores to synchronize threads is strictly DISALLOWED. Additionally, you are NOT PERMITTED to use any of the following: wait(), notify(), notifyAll(), the synchronized keyword (for methods or blocks), and any synchronized collections or synchronization tools that were not discussed in class.

You CAN, however, use the modifier volatile and the AtomicInteger and AtomicBoolean classes if you choose to.

Directions: Synchronize the students and teacher threads in the context of the problem described below. The number of students can be entered as a command line argument. Please refer to and read carefully the project notes, tips and guidelines before starting the project.